Update flight_model.cfg
Adjustment in drag (was way too draggy for a sleek plane). Indeed, when done so, the factory spec of cruising at FL180 at 370kph (200kts) was obtained with MP 30 (not max power, but max cruise power) when the parasite_drag_scalar was lowered to 0.5 (not just a nice number, it created perfect reference results).